Adding example to show labels in POD yaml (#14439)
* Adding example to show labels in POD yaml

This page talk about labels and its usage. Also it talk about syntax for specifying labels. But it does not show how we can have labels in resource manifest. So with this PR i am adding example which shows that how we can have labels in pod manifest.

* Changes to resolve review comments

Changes to resolve review comments