Revise “What is Kubernetes” (#19534)
* Expand list of places Kubernetes can run

As well as GKE, there are AKS, EKS, and more.

* Add page description

* Link to blog post about Borg

The blog post has (a better) link to Google's paper about Borg. Because
of the historical relevance of that paper and that blog article to the
history of Kubernetes, I think it's OK to link there.

Linking to the blog article provides a less academic introduction to the
history of Kubernetes, with the option to click through and learn things
in detail.