Reword and rename “Managing Compute Resources for Containers” (#18940)
* Reword resource limits section about ephemeral local storage

Co-Authored-By: Celeste Horgan <celeste@cncf.io>

* Move troubleshooting section to foot of page

It's OK to hyperlink to this section; in fact, I recommend it.

* Tweak section about monitoring key Pod resources

Monitoring reports memory use as well as CPU

* Fold 2 displayed headings into 1

Add a 2nd-level heading to cover two related headings, that I'm moving
to 3rd level. This change shortens the table of contents and still
leaves an entry there for readers to click on.

* Reword What's next section

* Revise punctuation

* Change page path to match title

Co-authored-by: Celeste Horgan <celeste@cncf.io>